requirements.txt
google-generativeai==0.3.2
streamlit==1.28.1
python-dotenv==1.0.0
langchain==0.0.350
langchain-google-genai==0.0.5
chromadb==0.4.18
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.0.3
python-multipart==0.0.6 
